requirements.txt
git+https://github.com/Pycord-Development/pycord
pymongo == 4.3.2
python-dotenv == 0.21.0